+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Build images with go 1.19
+- Add a `Transient` field to the task inputs
 
 ## [0.21.0] - 2022-08-01